fix(docs): autocomplete dx
