+++ apps/docs/content/components/autocomplete/custom-selector-icon.raw.jsx
@@ -1,6 +1,25 @@
 import {Autocomplete, AutocompleteItem} from "@nextui-org/react";
 
-import {SelectorIcon} from "./SelectorIcon";
+const SelectorIcon = (props) => (
+  <svg
+    aria-hidden="true"
+    fill="none"
+    focusable="false"
+    height="1em"
+    role="presentation"
+    stroke="currentColor"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+    strokeWidth="1.5"
+    viewBox="0 0 24 24"
+    width="1em"
+    {...props}
+  >
+    <path d="M0 0h24v24H0z" fill="none" stroke="none" />
+    <path d="M8 9l4 -4l4 4" />
+    <path d="M16 15l-4 4l-4 -4" />
+  </svg>
+);
 
 export const animals = [
   {label: "Cat", value: "cat", description: "The second most popular pet in the world"},
@@ -29,27 +48,6 @@ export const animals = [
   {label: "Otter", value: "otter", description: "A carnivorous mammal in the subfamily Lutrinae"},
   {label: "Crocodile", value: "crocodile", description: "A large semiaquatic reptile"},
 ];
-
-export const SelectorIcon = (props) => (
-  <svg
-    aria-hidden="true"
-    fill="none"
-    focusable="false"
-    height="1em"
-    role="presentation"
-    stroke="currentColor"
-    strokeLinecap="round"
-    strokeLinejoin="round"
-    strokeWidth="1.5"
-    viewBox="0 0 24 24"
-    width="1em"
-    {...props}
-  >
-    <path d="M0 0h24v24H0z" fill="none" stroke="none" />
-    <path d="M8 9l4 -4l4 4" />
-    <path d="M16 15l-4 4l-4 -4" />
-  </svg>
-);
 
 export default function App() {
   return (
