+++ apps/docs/content/components/autocomplete/dynamic.raw.jsx
@@ -27,6 +27,7 @@ export const animals = [
   {label: "Otter", value: "otter", description: "A carnivorous mammal in the subfamily Lutrinae"},
   {label: "Crocodile", value: "crocodile", description: "A large semiaquatic reptile"},
 ];
+
 export default function App() {
   return (
     <Autocomplete
